pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Projeto de Jupyter Notebook configurado no VSCode"
authors = ["douglaspands"]
readme = "README.md"
repository = "https://github.com/douglaspands/vscode-notebook"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^1.5.2"
selenium-wire = "^5.1.0"
pydash = "^5.1.2"
requests = "^2.28.1"
Pillow = "^9.4.0"
pytube = "^12.1.2"
ipykernel = "^6.20.0"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.4"
autoflake = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
makerequirements = "scripts.poetry:run_makerequirements"
format = "scripts.poetry:run_format"
pycacheremove = "scripts.poetry:run_pycache_remove"
